Revert "Get topic data handler (#15038)"
This reverts commit 85ef15c759c3726e95c3984239246d523eca1ad5.
