+++ ydb/core/http_proxy/ut/internal_counters.json
@@ -611,7 +611,7 @@
         "folder_id": "folder4",
         "database": "/Root"
         },
-        "value":243,
+        "value":238,
         "kind":"RATE"
     },
     {
