+++ ydb/core/viewer/json_pipe_req.cpp
@@ -544,48 +544,24 @@ void TViewerPipeClient::RequestBSControllerPDiskUpdateStatus(const NKikimrBlobSt
     SendRequestToPipe(pipeClient, request.Release());
 }
 
-THolder<NSchemeCache::TSchemeCacheNavigate> TViewerPipeClient::SchemeCacheNavigateRequestBuilder (
-        std::function<void (NSchemeCache::TSchemeCacheNavigate::TEntry&)> fillRequestEntryFunc
-) {
+void TViewerPipeClient::RequestSchemeCacheNavigate(const TString& path) {
     THolder<NSchemeCache::TSchemeCacheNavigate> request = MakeHolder<NSchemeCache::TSchemeCacheNavigate>();
     NSchemeCache::TSchemeCacheNavigate::TEntry entry;
+    entry.Path = SplitPath(path);
     entry.RedirectRequired = false;
     entry.Operation = NSchemeCache::TSchemeCacheNavigate::EOp::OpPath;
-    fillRequestEntryFunc(entry);
     request->ResultSet.emplace_back(entry);
-    return request;
-}
-
-void TViewerPipeClient::RequestSchemeCacheNavigateWithParams(const TString& path, ui32 access, bool showPrivate) {
-    auto request = SchemeCacheNavigateRequestBuilder(
-        [&](auto& entry) {
-            entry.Path = SplitPath(path);
-            entry.Access = access;
-            entry.ShowPrivatePath = showPrivate;
-        }
-    );
-    if (!Event->Get()->UserToken.empty())
-        request->UserToken = new NACLib::TUserToken(Event->Get()->UserToken);
-
-    SendRequest(MakeSchemeCacheID(), new TEvTxProxySchemeCache::TEvNavigateKeySet(request.Release()));
-}
-
-void TViewerPipeClient::RequestSchemeCacheNavigate(const TString& path) {
-    auto request = SchemeCacheNavigateRequestBuilder(
-        [&](auto& entry) {
-            entry.Path = SplitPath(path);
-        }
-    );
     SendRequest(MakeSchemeCacheID(), new TEvTxProxySchemeCache::TEvNavigateKeySet(request.Release()));
 }
 
 void TViewerPipeClient::RequestSchemeCacheNavigate(const TPathId& pathId) {
-    auto request = SchemeCacheNavigateRequestBuilder(
-        [&](auto& entry) {
-            entry.TableId.PathId = pathId;
-            entry.RequestType = NSchemeCache::TSchemeCacheNavigate::TEntry::ERequestType::ByTableId;
-        }
-    );
+    THolder<NSchemeCache::TSchemeCacheNavigate> request = MakeHolder<NSchemeCache::TSchemeCacheNavigate>();
+    NSchemeCache::TSchemeCacheNavigate::TEntry entry;
+    entry.TableId.PathId = pathId;
+    entry.RequestType = NSchemeCache::TSchemeCacheNavigate::TEntry::ERequestType::ByTableId;
+    entry.RedirectRequired = false;
+    entry.Operation = NSchemeCache::TSchemeCacheNavigate::EOp::OpPath;
+    request->ResultSet.emplace_back(entry);
     SendRequest(MakeSchemeCacheID(), new TEvTxProxySchemeCache::TEvNavigateKeySet(request.Release()));
 }
 
@@ -798,10 +774,6 @@ TString TViewerPipeClient::GetHTTPGATEWAYTIMEOUT(TString contentType, TString re
 
 TString TViewerPipeClient::GetHTTPBADREQUEST(TString contentType, TString response) {
     return Viewer->GetHTTPBADREQUEST(GetRequest(), std::move(contentType), std::move(response));
-}
-
-TString TViewerPipeClient::GetHTTPNOTFOUND(TString, TString) {
-    return Viewer->GetHTTPNOTFOUND(GetRequest());
 }
 
 TString TViewerPipeClient::GetHTTPINTERNALERROR(TString contentType, TString response) {
