+++ ydb/core/viewer/ya.make
@@ -72,7 +72,6 @@ SRCS(
     viewer_describe.h
     viewer_describe_topic.h
     viewer_feature_flags.h
-    viewer_topic_data.cpp
     viewer_graph.h
     viewer_healthcheck.h
     viewer_helper.h
